refactor(bt): split chip init into per-chip backend classes
Replace the scattered *_brcm_* module functions (each guarded by a
runtime /proc filesystem check) with a small BtChip strategy:
  - bt_chip.py: BtChip base (no-op hooks) + shared run/free_device
  - bt_brcm.py: BrcmChip (warm handoff, wake, power-off, re-warm)
  - bt_mtk.py:  MtkChip (module load + /dev/stpbt eviction)
  - bt_setup.py: make_chip() factory + memoized chip() accessor

transport/controller/daemon now call chip().on_transport_open() /
.power_off() / .ensure_powered() polymorphically; MTK inherits the
no-op defaults, so the chip is encoded in the type rather than a
per-call filesystem probe.

Behavior-preserving; validated on both Kindles (BRCM warm-handoff +
on/off/on cycle + icon, MTK module load + toggle), no new errors.
