+++ kindle_hid_passthrough/transport.py
@@ -74,10 +74,9 @@ async def create_bumble_device(transport_spec=None, configure=None):
         log.error(f"Transport open timed out after {config.transport_timeout}s")
         raise
 
-    # Broadcom (warm handoff): wake the chip now that the UART is open. No-op on
-    # other Kindles. Must be after open and before the first HCI command.
-    from bt_setup import wake_brcm_chip
-    wake_brcm_chip()
+    # chip hook: runs after the transport opens, before the first HCI command
+    from bt_setup import chip
+    chip().on_transport_open()
 
     try:
         device = Device.with_hci(
